relocated recording functions to a separate utils folder
